# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/shortlink/": {
            "get": {
                "description": "get a shorlink",
                "produces": [
                    "text/plain",
                    "application/json"
                ],
                "summary": "get a shortlink",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controller.ShortLink"
                            }
                        }
                    },
                    "404": {
                        "description": "NotFound",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "500": {
                        "description": "InternalServerError",
                        "schema": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "/api/v1/shortlink/{shortlink}": {
            "get": {
                "description": "get a shorlink",
                "produces": [
                    "text/plain",
                    "application/json"
                ],
                "summary": "get a shortlink",
                "parameters": [
                    {
                        "type": "string",
                        "example": "home",
                        "description": "the shortlink URL part (shortlink id)",
                        "name": "shortlink",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/controller.ShortLink"
                        }
                    },
                    "404": {
                        "description": "NotFound",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "500": {
                        "description": "InternalServerError",
                        "schema": {
                            "type": "integer"
                        }
                    }
                }
            },
            "put": {
                "description": "update a new shorlink",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain",
                    "application/json"
                ],
                "summary": "update existing shortlink",
                "parameters": [
                    {
                        "type": "string",
                        "example": "home",
                        "description": "the shortlink URL part (shortlink id)",
                        "name": "shortlink",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "shortlink spec",
                        "name": "spec",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1alpha1.ShortLinkSpec"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "404": {
                        "description": "NotFound",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "500": {
                        "description": "InternalServerError",
                        "schema": {
                            "type": "integer"
                        }
                    }
                }
            },
            "post": {
                "description": "create a new shorlink",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain",
                    "application/json"
                ],
                "summary": "create new shortlink",
                "parameters": [
                    {
                        "type": "string",
                        "example": "home",
                        "description": "the shortlink URL part (shortlink id)",
                        "name": "shortlink",
                        "in": "path"
                    },
                    {
                        "description": "shortlink spec",
                        "name": "spec",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1alpha1.ShortLinkSpec"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "301": {
                        "description": "MovedPermanently",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "302": {
                        "description": "Found",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "307": {
                        "description": "TemporaryRedirect",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "308": {
                        "description": "PermanentRedirect",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "404": {
                        "description": "NotFound",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "500": {
                        "description": "InternalServerError",
                        "schema": {
                            "type": "integer"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete shorlink",
                "produces": [
                    "text/plain",
                    "application/json"
                ],
                "summary": "delete shortlink",
                "parameters": [
                    {
                        "type": "string",
                        "example": "home",
                        "description": "the shortlink URL part (shortlink id)",
                        "name": "shortlink",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "404": {
                        "description": "NotFound",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "500": {
                        "description": "InternalServerError",
                        "schema": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "/{shortlink}": {
            "get": {
                "description": "redirect to target as per configuration of the shortlink",
                "produces": [
                    "text/html"
                ],
                "summary": "redirect to target",
                "parameters": [
                    {
                        "type": "string",
                        "description": "shortlink id",
                        "name": "shortlink",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "300": {
                        "description": "MultipleChoices",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "301": {
                        "description": "MovedPermanently",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "302": {
                        "description": "Found",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "303": {
                        "description": "SeeOther",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "304": {
                        "description": "NotModified",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "305": {
                        "description": "UseProxy",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "307": {
                        "description": "TemporaryRedirect",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "308": {
                        "description": "PermanentRedirect",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "404": {
                        "description": "NotFound",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "500": {
                        "description": "InternalServerError",
                        "schema": {
                            "type": "integer"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.ShortLink": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "spec": {
                    "$ref": "#/definitions/v1alpha1.ShortLinkSpec"
                },
                "status": {
                    "$ref": "#/definitions/v1alpha1.ShortLinkStatus"
                }
            }
        },
        "v1alpha1.ShortLinkSpec": {
            "type": "object",
            "properties": {
                "after": {
                    "description": "RedirectAfter specifies after how many seconds to redirect (Default=3)\n+kubebuilder:default:=3\n+kubebuilder:validation:Minimum=0\n+kubebuilder:validation:Maximum=99",
                    "type": "integer"
                },
                "code": {
                    "description": "Code is the URL Code used for the redirection.\nleave on default (200) when using the HTML behavior. However, if you whish to use a HTTP 3xx redirect, set to the appropriate 3xx status code\n+kubebuilder:validation:Enum=200;300;301;302;303;304;305;307;308\n+kubebuilder:default:=200",
                    "type": "integer",
                    "enum": [
                        200,
                        300,
                        301,
                        302,
                        303,
                        304,
                        305,
                        307,
                        308
                    ]
                },
                "target": {
                    "description": "Target specifies the target to which we will redirect\n+kubebuilder:validation:Required\n+kubebuilder:validation:MinLength=1",
                    "type": "string"
                }
            }
        },
        "v1alpha1.ShortLinkStatus": {
            "type": "object",
            "properties": {
                "count": {
                    "description": "Count represents how often this ShortLink has been called\n+kubebuilder:default:=0\n+kubebuilder:validation:Minimum=0",
                    "type": "integer"
                }
            }
        }
    }
}